third_party/WebKit/LayoutTests/http/tests/devtools/indexeddb/live-update-indexeddb-content-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that the IndexedDB database content live updates.

Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        Object store: objectStore1
            Index: index1
Object store marked needs refresh = false
Index marked needs refresh = false

Add entry to objectStore1:
Object store marked needs refresh = true
Index marked needs refresh = true
Dumping ObjectStore data:
    Object store: objectStore1
            Number of entries: 0
            Index: index1
                Number of entries: 0

Refresh views:
Object store marked needs refresh = false
Index marked needs refresh = false
Dumping ObjectStore data:
    Object store: objectStore1
            Number of entries: 1
            Key = testKey, value = [object Object]
            Index: index1
                Number of entries: 1
                Key = testKey, value = [object Object]

Delete entry from objectStore1:
Object store marked needs refresh = true
Index marked needs refresh = true
Dumping ObjectStore data:
    Object store: objectStore1
            Number of entries: 1
            Key = testKey, value = [object Object]
            Index: index1
                Number of entries: 1
                Key = testKey, value = [object Object]

Refresh views:
Object store marked needs refresh = false
Index marked needs refresh = false
Dumping ObjectStore data:
    Object store: objectStore1
            Number of entries: 0
            Index: index1
                Number of entries: 0
